refactor: Command 패키지 domain -> utils로 변경
